+++ src/sys/hw/mcu/esp32/_.rs
@@ -5,6 +5,7 @@
 
 crate::mods_in! {
     mod_ c3;
+    mod i2c;
     mod register;
     mod usb_serial_jtag;
 }
@@ -12,6 +13,7 @@ crate::mods_out! { // _mods
     _mods {
         pub use super::{
             c3::_all::*,
+            i2c::EspI2c,
             register::EspReg32,
             usb_serial_jtag::EspUsbSerialJtag,
         };